+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod ast;
-pub mod codegen;
 pub mod parser;
-pub mod pass;
+pub mod stage;
