+++ src/stage/codegen/register.rs
