+++ src/stage/type_pass/scopes.rs
